chore(charts): refactor for shared components and separate temp
