+++ src/components/ECharts/index.ts
@@ -9,6 +9,7 @@ export {
   formatUsdCompact,
 } from "./formatters"
 export { getChartWatermark } from "./options"
+export { CHART_PALETTE } from "./palette"
 export type {
   CategoryBarSeries,
   ChartMarkLine,
