+++ src/components/Profile/ProfilePage/components/PageSkeleton/index.tsx
@@ -1,7 +1,5 @@
 "use client"
 
-import { useEffect, useState } from "react"
-
 import { Box, Skeleton } from "@mui/material"
 
 import { pageCalcHeights } from "@/utils/constants"
@@ -13,35 +11,28 @@ const blockSx = {
   transform: "none" as const,
 }
 
-export const ProfilePageSkeleton = ({ isMobile }: ProfilePageSkeletonProps) => {
-  const [mounted, setMounted] = useState(false)
-  useEffect(() => setMounted(true), [])
+export const ProfilePageSkeleton = ({ isMobile }: ProfilePageSkeletonProps) => (
+  <Box
+    sx={{
+      width: "100%",
+      height: isMobile ? "auto" : `calc(100vh - ${pageCalcHeights.page})`,
+      overflowY: isMobile ? "visible" : "auto",
+      padding: isMobile ? "0" : "44px 44px 24px 44px",
+      display: "flex",
+      flexDirection: "column",
+      gap: "24px",
+    }}
+  >
+    <Skeleton sx={{ ...blockSx, height: "200px" }} />
 
-  if (!mounted) return null
-
-  return (
-    <Box
-      sx={{
-        width: "100%",
-        height: isMobile ? "auto" : `calc(100vh - ${pageCalcHeights.page})`,
-        overflowY: isMobile ? "visible" : "auto",
-        padding: isMobile ? "0" : "44px 44px 24px 44px",
-        display: "flex",
-        flexDirection: "column",
-        gap: "24px",
-      }}
-    >
-      <Skeleton sx={{ ...blockSx, height: "200px" }} />
-
-      <Box sx={{ display: "flex", flexDirection: "column", gap: "12px" }}>
-        <Skeleton sx={{ width: "180px", height: "20px" }} />
-        <Skeleton sx={{ ...blockSx, height: "320px" }} />
-      </Box>
-
-      <Box sx={{ display: "flex", flexDirection: "column", gap: "12px" }}>
-        <Skeleton sx={{ width: "200px", height: "20px" }} />
-        <Skeleton sx={{ ...blockSx, height: "200px" }} />
-      </Box>
+    <Box sx={{ display: "flex", flexDirection: "column", gap: "12px" }}>
+      <Skeleton sx={{ width: "180px", height: "20px" }} />
+      <Skeleton sx={{ ...blockSx, height: "320px" }} />
     </Box>
-  )
-}
+
+    <Box sx={{ display: "flex", flexDirection: "column", gap: "12px" }}>
+      <Skeleton sx={{ width: "200px", height: "20px" }} />
+      <Skeleton sx={{ ...blockSx, height: "200px" }} />
+    </Box>
+  </Box>
+)
